Refactor Managers
Refactor Managers
