--- Assets/Prefabs/Machineries.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e863d11a0e72f7549abb66aeb59cfd1b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
