--- Assets/Prefabs/Projectiles.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3908e5c0e4814a840b4a58cf131297d7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
