--- Assets/Prefabs/Train.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b455b5f0ad2c1a540bf78a9e9a72fb1d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
