+++ Assets/Scripts/Managers/TimeDown.cs
@@ -4,58 +4,58 @@ using UnityEngine.UI;
 [RequireComponent(typeof(Text))]
 public class TimeDown : MonoBehaviour
 {
-    [Header("Properties")]
-    [SerializeField]
-    private float timeRemaining = 300.0f;
-    [SerializeField]
-    private bool timerIsRunning = false;
+    //[Header("Properties")]
+    //[SerializeField]
+    //private float timeRemaining = 300.0f;
+    //[SerializeField]
+    //private bool timerIsRunning = false;
 
-    private Text textCountDown;
-    private GameManager gameManager;
+    //private Text textCountDown;
+    //private GameManager gameManager;
 
-    //Public functions
-    public void PauseTimer()
-    {
-        timerIsRunning = false;
-    }
+    ////Public functions
+    //public void PauseTimer()
+    //{
+    //    timerIsRunning = false;
+    //}
 
-    public void ResumeTimer()
-    {
-        timerIsRunning = true;
-    }
+    //public void ResumeTimer()
+    //{
+    //    timerIsRunning = true;
+    //}
 
-    //Private functions
-    private void Start()
-    {
-        textCountDown = GetComponent<Text>();
-        gameManager = GameManager.Instance;
-        timerIsRunning = true;
-    }
+    ////Private functions
+    //private void Start()
+    //{
+    //    textCountDown = GetComponent<Text>();
+    //    gameManager = GameManager.Instance;
+    //    timerIsRunning = true;
+    //}
 
-    private void Update()
-    {
-        if (timerIsRunning)
-        {
-            if (timeRemaining > 0f)
-            {
-                timeRemaining -= Time.deltaTime;
-                DisplayTime(timeRemaining);
+    //private void Update()
+    //{
+    //    if (timerIsRunning)
+    //    {
+    //        if (timeRemaining > 0f)
+    //        {
+    //            timeRemaining -= Time.deltaTime;
+    //            DisplayTime(timeRemaining);
 
-            }
-            else
-            {
-                timeRemaining = 0f;
-                timerIsRunning = false;
-                gameManager.YouWin();
-            }
+    //        }
+    //        else
+    //        {
+    //            timeRemaining = 0f;
+    //            timerIsRunning = false;
+    //            gameManager.YouWin();
+    //        }
 
-        }
-    }
+    //    }
+    //}
 
-    private void DisplayTime(float timeToDisplay)
-    {
-        float minutes = Mathf.FloorToInt(timeToDisplay / 60f);
-        float seconds = Mathf.FloorToInt(timeToDisplay % 60f);
-        textCountDown.text = string.Format("{0:00}:{1:00}", minutes, seconds);
-    }
+    //private void DisplayTime(float timeToDisplay)
+    //{
+    //    float minutes = Mathf.FloorToInt(timeToDisplay / 60f);
+    //    float seconds = Mathf.FloorToInt(timeToDisplay % 60f);
+    //    textCountDown.text = string.Format("{0:00}:{1:00}", minutes, seconds);
+    //}
 }
