--- Assets/Scripts/Train.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 914e20622df319540be8ad65bbed8924
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
