+++ Assets/Scripts/UI/FuelGaugeUI.cs
@@ -10,19 +10,19 @@ public class FuelGaugeUI : MonoBehaviour
     private float _updateSpeedSeconds = 0.5f;
 
     private Slider _sliderFuel;
-    private FuelController _fuelController;
+    //private FuelController _fuelController;
 
     private void Awake()
     {
         _sliderFuel = GetComponentInChildren<Slider>();
 
-        _fuelController = FindObjectOfType<FuelController>();
+       //_fuelController = FindObjectOfType<FuelController>();
 
-        if (_fuelController)
-        {
-            _fuelController.OnUpdateFuelUI += UpdateFuelUI;
-            _fuelController.OnFuelReloadUI += ReloadFuelUI;
-        }
+        //if (_fuelController)
+        //{
+        //    _fuelController.OnUpdateFuelUI += UpdateFuelUI;
+        //    _fuelController.OnFuelReloadUI += ReloadFuelUI;
+        //}
     }
 
     private void UpdateFuelUI(float pct)
